+++ apps/client/public/CHANGELOG.md
@@ -25,6 +25,7 @@ If you found some bugs or want us to add new feature, please do so via [our GitH
 
 ### 🩹 Fixes
 
+- Fixed `hideFiltersOnMapClick` preference not working
 - Photo section empty state icon opacity increased for better visibility
 
 # 2026-04-06